requirements.txt
absl-py==0.2.2
appdirs==1.4.3
apt-xapian-index==0.47
asn1crypto==0.24.0
astor==0.6.2
atomicwrites==1.1.5
attrs==18.1.0
backcall==0.1.0
beautifulsoup4==4.6.0
biocommons.seqrepo==0.3.9
biopython==1.69
bioutils==0.3.3
bleach==1.5.0
blinker==1.4
boto==2.48.0
boto3==1.7.42
botocore==1.10.42
brewer2mpl==1.4.1
bs4==0.0.1
bz2file==0.98
catfish==1.4.2
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.5.6
command-not-found==0.3
configparser==3.5.0
cryptography==2.2.2
cycler==0.10.0
Cython==0.28.3
dask==0.19.2
datacache==1.0.0
decorator==4.3.0
defer==1.0.6
devscripts==1.0.0
docutils==0.14
entrypoints==0.2.3
enum34==1.1.6
et-xmlfile==1.0.1
fail2ban==0.9.3
fastcluster==1.1.25
Flask==1.0.2
gast==0.2.0
gensim==3.4.0
ggplot==0.11.5
Glances==3.0.2
graphviz==0.8.3
grpcio==1.12.1
gtfparse==1.0.7
h5py==2.8.0
hdbscan==0.8.13
hgvs==1.2.4
html5lib==0.9999999
httplib2==0.11.3
idna==2.7
iniherit==0.3.9
ipykernel==4.8.2
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
jdcal==1.4
jedi==0.12.0
jeepney==0.3.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keyring==13.0.0
kiwisolver==1.0.1
language-selector==0.1
lifelines==0.14.4
lightdm-gtk-greeter-settings==1.2.1
lxml==4.2.1
Magic-file-extensions==0.2
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
memoized-property==1.0.3
menulibre==2.1.3
mistune==0.8.3
mock==2.0.0
more-itertools==4.2.0
mugshot==0.3.1
MulticoreTSNE==0.1
mysql-connector-python==2.0.4
mysqlclient==1.3.7
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.3
notebook==5.5.0
numexpr==2.6.5
numpy==1.14.5
nvidia-ml-py3==7.352.0
oauthlib==2.1.0
onboard==1.2.0
oneconf==0.3.9
opencv-python==3.4.1.15
openpyxl==2.5.4
openslide-python==1.1.1
pandas==0.23.1
pandocfilters==1.4.2
Parsley==1.3
parso==0.2.1
patsy==0.5.0
pbr==4.0.4
pep8==1.7.1
pexpect==4.6.0
PhenoGraph==1.5.2
pickleshare==0.7.4
Pillow==5.2.0
pip-autoremove==0.9.1
pipdate==0.3.1
piston-mini-client==0.7.5
pluggy==0.6.0
progressbar33==2.4
prompt-toolkit==1.0.15
protobuf==3.6.0
psutil==5.4.6
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.5.2
py==1.5.3
pyasn1==0.4.3
pycairo==1.17.0
pycparser==2.18
pycrypto==2.6.1
pycuda==2017.1.1
pycups==1.9.73
pycurl==7.43.0.2
pydot==1.2.4
pyensembl==1.2.6
Pygments==2.2.0
pygobject==3.20.0
pyinotify==0.9.6
PyJWT==1.6.4
pylibdmtx==0.1.7
PyMySQL==0.8.1
pyodbc==4.0.23
pyOpenSSL==18.0.0
pyparsing==2.2.0
pysam==0.14.1
pytesseract==0.2.4
pytest==3.6.2
python-apt==1.1.0b1+ubuntu0.16.4.2
python-dateutil==2.7.3
python-debian==0.1.32
python-systemd==231
pytools==2018.4
pytz==2018.4
PyVCF==0.6.8
PyWavelets==1.0.0
pyxdg==0.26
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
reportlab==3.4.0
requests==2.19.1
roman==3.0
s3transfer==0.1.13
scikit-allel==1.1.10
scikit-image==0.14.0
scikit-learn==0.19.1
scipy==1.1.0
screen-resolution-extra==0.0.0
seaborn==0.9.0
SecretStorage==3.0.1
Send2Trash==1.5.0
serializable==0.1.1
sessioninstaller==0.0.0
Shapely==1.6.4.post1
simplegeneric==0.8.1
simplejson==3.15.0
six==1.11.0
sklearn==0.0
smart-open==1.5.7
software-center-aptd-plugins==0.0.0
sparse==0.3.1
sphinx-rtd-theme==0.4.0
SQLAlchemy==1.2.8
ssh-import-id==5.6
statsmodels==0.9.0
tables==3.4.4
tensorboard==1.11.0
tensorflow-gpu==1.11.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
text-unidecode==1.2
tflearn==0.3.2
tinytimer==0.0.0
toolz==0.9.0
tornado==5.0.2
tqdm==4.23.4
traitlets==4.3.2
typechecks==0.1.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unicodecsv==0.14.1
urllib3==1.23
virtualenv==16.0.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.2.1
xkit==0.0.0
xlrd==1.1.0
xmltodict==0.11.0
yoyo-migrations==5.1.5
